# packages/contracts/openapi/sharedhouse-v1.yaml
openapi: 3.1.0
info:
  title: SharedHouse API
  version: 1.3.0
  description: Versioned public contract for the SharedHouse coordination platform.
servers:
  - url: https://houseapi.dohotstudio.com
    description: SharedHouse production API
  - url: http://localhost:3000
    description: Synthetic local development environment
...
paths:
...
  /v1/auth/register:
    post:
      operationId: registerAccount
      summary: Register an adult account and start email verification
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '202':
          description: Registration was accepted. The response is intentionally generic.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationAccepted'
        '400':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /v1/auth/verify-email:
    post:
      operationId: verifyEmail
      summary: Consume a single-use email verification code and create a session
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified and session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /v1/auth/resend-verification:
    post:
      operationId: resendEmailVerification
      summary: Request another single-use email verification code
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendEmailVerificationRequest'
      responses:
        '202':
          description: The request was accepted without disclosing account existence.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationAccepted'
        '400':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /v1/auth/sign-in:
    post:
      operationId: signIn
      summary: Create a session with email and password
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
      responses:
        '200':
          description: Session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /v1/auth/refresh:
    post:
      operationId: refreshSession
      summary: Rotate a refresh token and issue a new session token pair
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshSessionRequest'
      responses:
        '200':
          description: Session token pair rotated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /v1/auth/sign-out:
    post:
      operationId: signOut
      summary: Revoke the current session
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Current session revoked.
        '401':
          $ref: '#/components/responses/Problem'
  /v1/account:
    get:
      operationId: getAccount
      summary: Return the authenticated account
      tags: [Account]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Safe account representation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSummary'
        '401':
          $ref: '#/components/responses/Problem'
...
  /v1/households/{householdId}/invitations:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
    get:
      operationId: listHouseholdInvitations
      summary: List safe invitation metadata for household administrators
      tags: [Invitations]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Invitations visible to an authorised household administrator.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HouseholdInvitationSummary'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
    post:
      operationId: createHouseholdInvitation
      summary: Create a seven-day single-use household invitation
      tags: [Invitations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHouseholdInvitationRequest'
      responses:
        '201':
          description: Invitation created; the secret token is returned only once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdInvitationCreated'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /v1/households/{householdId}/invitations/{invitationId}:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - $ref: '#/components/parameters/InvitationId'
    delete:
      operationId: revokeHouseholdInvitation
      summary: Revoke a pending household invitation idempotently
      tags: [Invitations]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Invitation revoked or already terminal without acceptance.
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /v1/invitations/{invitationToken}:
    parameters:
      - $ref: '#/components/parameters/InvitationToken'
    get:
      operationId: previewHouseholdInvitation
      summary: Preview safe household invitation metadata
      tags: [Invitations]
      responses:
        '200':
          description: Safe invitation preview without the restricted email address.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdInvitationPreview'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
  /v1/invitations/{invitationToken}/accept:
    parameters:
      - $ref: '#/components/parameters/InvitationToken'
    post:
      operationId: acceptHouseholdInvitation
      summary: Accept a single-use invitation as the authenticated account
      tags: [Invitations]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Invitation accepted or safely replayed by the accepting account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptHouseholdInvitationResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Opaque
  parameters:
...
    HouseholdId:
      name: householdId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    InvitationId:
      name: invitationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InvitationToken:
      name: invitationToken
      in: path
      required: true
      schema:
        type: string
        pattern: '^sh_inv_[A-Za-z0-9_-]{43}$'
        writeOnly: true
  responses:
    Problem:
      description: Request failed with a safe problem-details response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, code, correlationId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
        correlationId:
          type: string
          format: uuid
        detail:
          type: string
        violations:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
    AccountSummary:
      type: object
      additionalProperties: false
      required: [id, email, emailVerified, displayName, preferredLocale]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        displayName:
          type: string
        preferredLocale:
          type: string
          enum: [en, ro]
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
        - displayName
        - preferredLocale
        - ageConfirmed
        - termsAccepted
        - marketingConsent
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 15
          maxLength: 128
          writeOnly: true
        displayName:
          type: string
          minLength: 1
          maxLength: 80
        preferredLocale:
          type: string
          enum: [en, ro]
        ageConfirmed:
          type: boolean
          const: true
        termsAccepted:
          type: boolean
          const: true
        marketingConsent:
          type: boolean
    RegistrationAccepted:
      type: object
      additionalProperties: false
      required: [verificationRequired]
      properties:
        verificationRequired:
          type: boolean
          const: true
        developmentVerificationCode:
          type: string
          pattern: '^[0-9]{8}$'
          description: Present only in explicitly non-production local or test environments.
    VerifyEmailRequest:
      type: object
      additionalProperties: false
      required: [email, code]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        code:
          type: string
          pattern: '^[0-9]{8}$'
          writeOnly: true
        deviceName:
          type: string
          minLength: 1
          maxLength: 80
    ResendEmailVerificationRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
    SignInRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 128
          writeOnly: true
        deviceName:
          type: string
          minLength: 1
          maxLength: 80
    RefreshSessionRequest:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 40
          maxLength: 256
          writeOnly: true
    SessionResponse:
      type: object
      additionalProperties: false
      required:
        - accessToken
        - refreshToken
        - accessTokenExpiresAt
        - refreshTokenExpiresAt
        - account
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        accessTokenExpiresAt:
          type: string
          format: date-time
        refreshTokenExpiresAt:
          type: string
          format: date-time
        account:
          $ref: '#/components/schemas/AccountSummary'
    HouseholdConfiguration:
      type: object
      additionalProperties: false
      required: [name, countryCode, timezone, currency, firstDayOfWeek, cycleType, cycleAnchor]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        countryCode:
          type: string
          pattern: '^[A-Z]{2}$'
        timezone:
          type: string
          minLength: 1
          maxLength: 64
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        firstDayOfWeek:
          type: integer
          enum: [1, 6, 7]
          description: ISO weekday number for Monday, Saturday or Sunday.
        cycleType:
          type: string
          enum: [weekly, fourteen_day, calendar_month]
        cycleAnchor:
          type: string
          format: date
    HouseholdSummary:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/HouseholdConfiguration'
        - type: object
          required: [id, role, status, version, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            role:
              type: string
              enum: [owner, admin, member, read_only]
            status:
              type: string
              const: active
            version:
              type: integer
              minimum: 1
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    CreateHouseholdInvitationRequest:
      type: object
      additionalProperties: false
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, member, read_only]
        email:
          type: [string, 'null']
          format: email
          maxLength: 254
          description: Optional account-email restriction. Never exposed by public preview.
    HouseholdInvitationSummary:
      type: object
      additionalProperties: false
      required: [id, householdId, householdName, role, email, status, expiresAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        householdName:
          type: string
          minLength: 1
          maxLength: 100
        role:
          type: string
          enum: [admin, member, read_only]
        email:
          type: [string, 'null']
          format: email
        status:
          type: string
          enum: [pending, accepted, revoked, expired]
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    HouseholdInvitationCreated:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/HouseholdInvitationSummary'
        - type: object
          required: [token]
          properties:
            token:
              type: string
              pattern: '^sh_inv_[A-Za-z0-9_-]{43}$'
              writeOnly: true
    HouseholdInvitationPreview:
      type: object
      additionalProperties: false
      required: [householdName, role, emailRestricted, status, expiresAt]
      properties:
        householdName:
          type: string
          minLength: 1
          maxLength: 100
        role:
          type: string
          enum: [admin, member, read_only]
        emailRestricted:
          type: boolean
        status:
          type: string
          enum: [pending, expired, unavailable]
        expiresAt:
          type: string
          format: date-time
    AcceptHouseholdInvitationResponse:
      type: object
      additionalProperties: false
      required: [household]
      properties:
        household:
          $ref: '#/components/schemas/HouseholdSummary'
...